Moving OG context to OG core.
