--- modules/og_context/config/install/og_context.ogcontextneogation.yml
@@ -1,1 +1,0 @@
-og_context:
