--- modules/og_context/og_context.services.yml
@@ -1,4 +1,0 @@
-services:
-  plugin.manager.og.context:
-    class: Drupal\og_context\Plugin\OgContextManager
-    parent: default_plugin_manager
