--- modules/og_context/src/Entity/OgContextConfigInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Drupal\og_context\Entity;
-
-use Drupal\Core\Config\Entity\ConfigEntityInterface;
-
-/**
- * Provides an interface for defining OG context config entities.
- */
-interface OgContextConfigInterface extends ConfigEntityInterface {
-
-  // Add get/set methods for your configuration properties here.
-}
