--- modules/og_context/src/Form/OgContextNeogation.php
@@ -1,58 +1,0 @@
-<?php
-
-namespace Drupal\og_context\Form;
-
-use Drupal\Core\Form\FormBase;
-use Drupal\Core\Form\FormStateInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Drupal\Core\Entity\EntityManager;
-
-/**
- * Class OgContextNeogation.
- *
- * @package Drupal\og_context\Form
- */
-class OgContextNeogation extends FormBase {
-
-  /**
-   * Drupal\Core\Entity\EntityManager definition.
-   *
-   * @var Drupal\Core\Entity\EntityManager
-   */
-  protected $entityManager;
-  public function __construct(
-    EntityManager $entity_manager
-  ) {
-    $this->entityManager = $entity_manager;
-  }
-
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('entity.manager')
-    );
-  }
-
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getFormId() {
-    return 'og_context_neogation';
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function buildForm(array $form, FormStateInterface $form_state) {
-
-    return $form;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function submitForm(array &$form, FormStateInterface $form_state) {
-
-  }
-
-}
