--- modules/og_context/src/Plugin/OgContext/Enitty.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Drupal\og_context\Plugin\OgContext;
-
-use Drupal\og_context\Plugin\OgContextBase;
-
-/**
- * @OgContext(
- *  id = "entity",
- *  title = "Entity base",
- *  description = @Translation("Get the group from the current entity.")
- * )
- */
-class Entity extends OgContextBase {
-
-}
