+++ src/OgContextNegotiationInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @file
+ * Contains Drupal\og\OgContextNegotiationInterface.
+ */
+
+namespace Drupal\og;
+
+use Drupal\Core\Config\Entity\ConfigEntityInterface;
+
+/**
+ * Provides an interface for defining OG context negotiation entities.
+ */
+interface OgContextNegotiationInterface extends ConfigEntityInterface {
+
+}
